Keys for data formed, abstracing front end functions for clearer read
