+++ src/controllers/socketHandlers/game.handler.ts
@@ -1,6 +1,7 @@
 import { Server, Socket } from 'socket.io'
 import { GameData } from 'redis/game'
 import { EVENTS } from 'sockets/game.sockets'
+import { Player } from 'database/models'
 export const joinRoom = async(io: Server, socket: Socket, username: string, roomCode: string, sessionToken: string) => {
     // maybe insta search for player and then return somehting like wrong player and room combo
     if (!await GameData.roomExists(roomCode)){
@@ -13,7 +14,6 @@ export const joinRoom = async(io: Server, socket: Socket, username: string, room
         // Perhaps wrap validation fucntions in one function and one redis calls instead of 3
         // but due to relativly low amout of requests this will work in this case as well since redis is really fast
         // if this is making problems wrap functions in 1 and do 1 db call
-        console.log(sessionToken)
         if (!await room.playerExists(username) && !await room.checkSessionToken(username, sessionToken)) {
             socket.emit(EVENTS.JOIN_ROOM, {
                 MSG: 'Igrac nije registrovan',
@@ -21,8 +21,13 @@ export const joinRoom = async(io: Server, socket: Socket, username: string, room
             })
         } else {
             const data = await room.retrieveJoinRoomData(username)
+            await socket.join(roomCode)
             socket.emit(EVENTS.JOIN_ROOM, {
                 ...data
+            })
+            socket.to(roomCode).emit(EVENTS.PLAYER_JOINED, {
+                username,
+                points: data.points
             })
         }
     }
