+++ src/validators/socketValidator.ts
@@ -1,5 +1,6 @@
 import { Server, Socket } from 'socket.io'
 import { EVENTS } from 'sockets/game.sockets'
+import { UsernameRegEx, RoomCodeRegEx, SessionTokenRegEx} from 'utils/strings'
 export const joinRoomValidator = async(
         io: Server,
         socket: Socket,
@@ -7,9 +8,9 @@ export const joinRoomValidator = async(
         roomCode: string,
         sessionToken: string) => {
     try {
-        const roomReg = await new RegExp('^[A-Za-z0-9]{8}$','g').test(roomCode)
-        const usernameReg = await new RegExp('^[A-Za-z0-9а-шА-ШčČćĆžŽšŠđĐђјљњћџЂЈЉЊЋЏ ]{4,16}$','g').test(username)
-        const tokenReg = await new RegExp('^[A-Za-z0-9]{96}$','g').test(sessionToken)
+        const roomReg = await new RegExp(RoomCodeRegEx,'g').test(roomCode)
+        const usernameReg = await new RegExp(UsernameRegEx,'g').test(username)
+        const tokenReg = await new RegExp(SessionTokenRegEx,'g').test(sessionToken)
         if(roomReg && usernameReg && tokenReg) return true
         return false
     } catch(e) {
